Move to errors.ts
