+++ src/shadowbox/model/errors.ts
